Branch2b: Transitions Fully Added
Added in all the transitions, alongside made the cards tighter together for all levels. Hopefully they're uniform enough for everyone involved. Level 2 is the only one that is left non uniform due to its pillar-like transition state.

Note: This is a branch that alters all the fly points to include all the transitions. Refer to Branch2a for rollback if lead or customer do not like.