Branch5b: Particle Update
Finally managed to get the particles to glow and have a nice sheen to them behind the test cards. Need to have George come in with me so we can workshop a way to get the cards to have this particle appear and disappear after certain actions. (Selecting/Hovering, etc)